new updates on restructuring & reorganization
